analytical_module/requirements.txt
joblib
google-cloud-bigquery
google-cloud-storage
pandas
numpy
pmdarima
matplotlib
db-dtypes
